+++ hw-02_huffman/include/writer.h
@@ -8,7 +8,7 @@ namespace DataProcessing {
         explicit Writer(std::ostream& stream);
         ~Writer();
 
-        void write(uint8_t msg, size_t msg_size);
+        void write(uint16_t msg, size_t msg_size);
         void writeInt(uint32_t number);
 
         int32_t bytesInserted();
@@ -18,9 +18,9 @@ namespace DataProcessing {
         void writeByte();
 
         size_t buffer_size_ = 0;
-        uint16_t buffer_ = 0;
+        uint32_t buffer_ = 0;
         std::ostream& stream_;
     };
 
-    uint8_t bit_mask(size_t length = CHAR_BIT);
+    uint16_t bit_mask(size_t length = CHAR_BIT);
 }
