+++ hw-02_huffman/src/code.cpp
@@ -1,7 +1,7 @@
 #include "code.h"
 
 namespace Huffman {
-    Code::Code(uint8_t symbol, size_t length, uint8_t code): symbol_(symbol), length_(length), code_(code) {}
+    Code::Code(uint8_t symbol, size_t length, uint16_t code): symbol_(symbol), length_(length), code_(code) {}
 
     bool Code::operator<(const Code &a) const {
         return symbol_ < a.symbol_;
